+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to
 
 ### Fixed
 
+- 🐛(backend) fix flaky test with search contact #605
 - 🐛(backend) fix flaky test with team access #646
 - 🧑‍💻(dimail) remove 'NoneType: None' log in debug mode
 - 🐛(frontend) fix flaky e2e test #636